requirements.txt
arabic-reshaper
asgiref
asn1crypto
certifi
cffi
chardet
charset-normalizer
click
colorama
crispy-bootstrap5
cryptography
cssselect2
defusedxml
dj-database-url
Django
django-allauth
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-environ
django-filter
django-rest-auth
django-rest-framework
djangorestframework
djangorestframework-simplejwt
et-xmlfile
h11
h2
hpack
hstspreload
html5lib
httpcore
httpx
hyperframe
idna
lxml
oauthlib
oscrypto
pillow
pycparser
pyHanko
pyhanko-certvalidator
PyJWT
pypdf
pypng
python-bidi
python-dotenv
python3-openid
pytz
PyYAML
qrcode
reportlab
requests
requests-oauthlib
rfc3986
six
sniffio
sqlparse
svglib
tinycss2
typing_extensions
tzdata
tzlocal
uritools
urllib3
webencodings
xhtml2pdf
XlsxWriter